Complete boundary conditional feature
