+++ tests/Feature/Patterns/files/Conditional_BoundaryEvent_CallActivity_Interrupting.json
@@ -6,9 +6,18 @@
         ],
         "result": [
             "_6",
-            "_20",
-            "_4",
             "_9"
+        ]
+    },
+    {
+        "startEvent": "StartEvent_1",
+        "data": {"b":1},
+        "events": [
+        ],
+        "result": [
+            "_6",
+            "_20",
+            "_8"
         ]
     }
 ]
